fix: pricings test suite & improved documentation
